Update planogram ID type, DI resolvers, and mapping profiles
- Changed GetPlanogramShelves parameter from int to long to support larger IDs.
- Registered mapping resolvers in DI for shelf/part mapping.
- Updated AutoMapper registration to use AddMaps for all profiles.
- Switched mapping from CountryList to CountriesList in profiles.
- Cleaned up and added necessary using directives.

+++ Controllers/planm/EditPlanApiController.cs
@@ -300,7 +300,7 @@ namespace PlanMatr_API.Controllers.planm
 
         [Route("api/v2/planx/get-planogram-shelves/{planogramId}")]
         [HttpGet]
-        public async Task<IActionResult> GetPlanogramShelves(int planogramId)
+        public async Task<IActionResult> GetPlanogramShelves(long planogramId)
         {
             try
             {
